+++ app/controllers/janela/panes_controller.rb
@@ -1,15 +1,15 @@
 module Janela
-  class VisualsController < ApplicationController
+  class PanesController < ApplicationController
     def show
-      @visual = Visual.new(
+      @pane = Pane.new(
         definition: Janela.definition!(params.require(:model)),
         measure: params.require(:measure).to_sym,
-        dimension: params.require(:by).to_sym,
+        dimension: params[:dimension]&.to_sym,
         renderer: params.fetch(:as, "table"),
         filters: filters
       )
 
-      @result = @visual.result(on: janela_scope(@visual.model))
+      @result = @pane.result(on: janela_scope(@pane.model))
     end
 
     private