+++ test/dummy/config/routes.rb
@@ -1,4 +1,4 @@
 Rails.application.routes.draw do
   root "dashboards#show"
-  mount Janela::Engine => "/janela"
+  mount Janela::Engine => "/dashboards"
 end