+++ test/janela_test.rb
@@ -7,6 +7,6 @@ class JanelaTest < ActiveSupport::TestCase
 
   test "the engine is mounted in the host application" do
     Rails.application.reload_routes!
-    assert_equal "/janela", Rails.application.routes.url_helpers.janela_path
+    assert_equal "/dashboards", Rails.application.routes.url_helpers.janela_path
   end
 end